[front] refactor: split attribution breakdown
